apps/demo-be-python-fastapi/pyproject.toml
[project]
name = "demo-be-python-fastapi"
version = "0.1.0"
description = "Python/FastAPI demo backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "pyjwt>=2.9",
    "bcrypt>=4.0",
    "pydantic[email]>=2.9",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-bdd>=7.3",
    "coverage[toml]>=7.6",
    "httpx>=0.27",
    "ruff>=0.8",
    "pyright>=1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/demo_be_python_fastapi"]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-bdd>=7.3",
    "coverage[toml]>=7.6",
    "httpx>=0.27",
    "ruff>=0.8",
    "pyright>=1.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_steps.py"]
markers = [
    "unit: unit tests",
    "integration: BDD integration tests",
]

[tool.coverage.run]
source = ["demo_be_python_fastapi"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true

[tool.coverage.lcov]
output = "coverage/lcov.info"

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "PT"]
ignore = [
    "B008",  # FastAPI Depends() in argument defaults is idiomatic and correct
]

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.13"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"
